a_clean_compress/requirements.txt
# Core data processing
pandas>=1.5.0,<2.0.0
datacleaner>=0.1.5

# HTTP requests for API interactions
requests>=2.31.0,<3.0.0

# Note: lzma, gzip, and bz2 are built-in Python libraries, no installation needed

# Command line interface
click>=8.1.0,<9.0.0

# Environment variable management
python-dotenv>=1.0.0,<2.0.0

# Vector similarity search for RAG implementation
sentence-transformers>=2.2.0,<3.0.0

# Enhanced CLI features
click-help-colors>=0.9.1,<1.0.0
click-option-group>=0.5.3,<1.0.0

# Torrent handling
qbittorrent-api>=2024.1.1
python-magic>=0.4.27 